Added link to bash.md reference in documentation
Added second sentence and link to this paragraph: 

"The other build step runs a Bash script as the build step using the built in SSH Exec runner. You can see how to do that in teamcityloadimpacy/TC_bash.md."

We're going to use the Powershell doc on the marketing site's landing page, so I want to have a reference to the Bash resource on here, as well.